cmd/pressure-vessel: come out of alarm if safe
The program will come out of alarm state if the pressure is below max and 30 seconds has elapsed. This is done because we don't think it's
actually pumping for that long and may be entering alarm state unnecessarily. We need to get it working so this should safely allow the
software to keep pumping, giving the pump 30 second rests if it's actually pumping for max pump time.
Initial pressure values are also discarded because there is initial noise.